+++ test/test_astnode.py
@@ -45,7 +45,7 @@ def test_bipabstractcitem00():
     # equality
     assert id(aci) != id(hxf.root_node)
     assert aci == hxf.root_node
-    assert aci != hxf.root_node.stmt_childs[0]
+    assert aci != hxf.root_node.stmt_children[0]
     assert aci.__eq__(0x10) == NotImplemented
     assert aci.__ne__(0x10) == NotImplemented
     assert aci != 0x10
@@ -59,7 +59,7 @@ def test_bipcnode00():
     assert not cn.is_expr
     gentst_cnode(cn)
     gentst_cnodestmt(cn)
-    cnc = cn.stmt_childs[0] # first child, this should be a CNodeStmtExpr
+    cnc = cn.stmt_children[0] # first child, this should be a CNodeStmtExpr
     assert isinstance(cnc, CNodeStmtExpr)
     assert cnc.is_statement
     assert not cnc.is_expr
@@ -86,10 +86,10 @@ def test_bipcnode00():
     assert len(cna.ops) == 2
     assert isinstance(cna.find_final_left_node(), CNodeExprVar)
     # cnodeStmt
-    assert len(cn.stmt_childs) != 0
-    assert len(cnc.stmt_childs) == 0
-    assert len(cn.expr_childs) == 0
-    assert len(cnc.expr_childs) == 1
+    assert len(cn.stmt_children) != 0
+    assert len(cnc.stmt_children) == 0
+    assert len(cn.expr_children) == 0
+    assert len(cnc.expr_children) == 1
     hxf2 = HxCFunc.from_addr(0x0180002524)
     assert isinstance(hxf2.get_cnode_label(6), CNode)
     assert hxf2.get_cnode_label(42) is None
@@ -105,7 +105,7 @@ def test_biphxcitem00():
     hxf = HxCFunc.from_addr(0x01800D2FF0)
     hi = hxf.hx_root_stmt
     gentst_hxcstmt(hi)
-    hic = hi.stmt_childs[0] # first child, this should be a CNodeStmtExpr
+    hic = hi.stmt_children[0] # first child, this should be a CNodeStmtExpr
     assert isinstance(hic, HxCStmtExpr)
     gentst_hxcstmt(hic)
     hia = hic.value # first asg